Check `isMainQueue` instead of just `isMainThread` since there's no guarantee that the main thread is running the main queue
The main queue always runs on the main thread. But the main thread can also run a background queue. That's because the Dispatch framework tries to reuse threads as much as possible. So if the main thread is idle, it might as well be used to run a background queue.